fix: fixed build error on swift linux env /app/Source/SwiftyJSON/SwiftyJSON.swift:1211:51: error: ambiguous use of operator '<' 1209 | 1210 |     switch (lhs.type, rhs.type) { 1211 |     case (.number, .number): return lhs.rawNumber < rhs.rawNumber      |                                                   `- error: ambiguous use of operator '<' 1212 |     case (.string, .string): return lhs.rawString < rhs.rawString 1213 |     default:                 return false      : 1245 | } 1246 | 1247 | func < (lhs: NSNumber, rhs: NSNumber) -> Bool {      |      `- note: found this candidate 1248 | 1249 |     switch (lhs.isBool, rhs.isBool) {
Foundation.NSNumber.<:2:20: note: found this candidate in module 'Foundation'

+++ Source/SwiftyJSON/SwiftyJSON.swift
@@ -1173,7 +1173,7 @@ public func == (lhs: JSON, rhs: JSON) -> Bool {
 public func <= (lhs: JSON, rhs: JSON) -> Bool {
 
     switch (lhs.type, rhs.type) {
-    case (.number, .number): return lhs.rawNumber <= rhs.rawNumber
+    case (.number, .number): return lhs.rawNumber.doubleValue <= rhs.rawNumber.doubleValue
     case (.string, .string): return lhs.rawString <= rhs.rawString
     case (.bool, .bool):     return lhs.rawBool == rhs.rawBool
     case (.array, .array):   return lhs.rawArray as NSArray == rhs.rawArray as NSArray
@@ -1208,7 +1208,7 @@ public func > (lhs: JSON, rhs: JSON) -> Bool {
 public func < (lhs: JSON, rhs: JSON) -> Bool {
 
     switch (lhs.type, rhs.type) {
-    case (.number, .number): return lhs.rawNumber < rhs.rawNumber
+    case (.number, .number): return lhs.rawNumber.doubleValue < rhs.rawNumber.doubleValue
     case (.string, .string): return lhs.rawString < rhs.rawString
     default:                 return false
     }
